--- src/main/java/inu/codin/codin/domain/elasticsearch/event/LectureDeletedEvent.java
@@ -1,3 +1,0 @@
-package inu.codin.codin.domain.elasticsearch.event;
-
-public record LectureDeletedEvent(Long lectureId) { }
