+++ src/main/java/inu/codin/codin/domain/lecture/document/dto/SemesterInfo.java
@@ -1,4 +1,4 @@
-package inu.codin.codin.domain.elasticsearch.document.dto;
+package inu.codin.codin.domain.lecture.document.dto;
 
 import inu.codin.codin.domain.lecture.entity.LectureSemester;
 import inu.codin.codin.domain.lecture.entity.Semester;
