+++ src/main/java/inu/codin/codin/domain/lecture/event/LectureDeletedEvent.java
@@ -1,0 +1,4 @@
+package inu.codin.codin.domain.lecture.event;
+
+public record LectureDeletedEvent(Long lectureId) {
+}
