+++ src/main/java/inu/codin/codin/domain/lecture/repository/jpa/LectureRepository.java
@@ -1,4 +1,4 @@
-package inu.codin.codin.domain.lecture.repository;
+package inu.codin.codin.domain.lecture.repository.jpa;
 
 import inu.codin.codin.domain.lecture.entity.Lecture;
 import org.springframework.data.domain.Pageable;
